pythonapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
opensearch-py==3.0.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
